refactor(api): move inline ninja Schemas to schemas.py (citation, media, provenance)
Citation, media, and provenance had request/response Schemas defined directly
in api.py. Pulled them out into each app's schemas.py (created for citation;
appended for media and provenance). Drops baselines in
test_openapi_boundaries.py to 0 for those three apps. Accounts (4 inline
schemas in a 346-line file with no schemas.py) left as-is.
